Create README.md
A common problem that is encountered while training machine learning models is imbalanced data. An imbalanced datasets can lead to inaccurate results even when best models are used to process that data. If the data is biased, the results will also be biased. If a data set is biased towards one class, the algorithm will also be biased towards the same class

In classification techniques, we would have tons of data points and label pairs. Labels are the class associated with each data point. If the distribution of the labels is not moderately uniform, then the datasets is called imbalanced.

Challenges faced with Imbalanced datasets
ML algorithms misleading Accuracy because of the unequal distribution in dependent variable.
This causes the performance of existing classifiers to get biased towards majority class.

Sampling based approaches to handling Imbalanced Data
Oversampling, by adding more of the minority class so it has more effect on the machine learning algorithm
Under-sampling, by removing some of the majority class so it has less effect on the machine learning algorithm
Hybrid, a mix of oversampling and under-sampling


https://medium.com/@pradeep.dhote9/different-way-to-handle-imbalanced-data-set-in-ml-8c960f515a34